rollback metadata changes in meta.c

+++ src/meta.c
@@ -847,7 +847,7 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
     }
 #endif
 
-    int32_t fileSizes[FTI_BUFS];
+    long fileSizes[FTI_BUFS];
     MPI_Allgather(&FTI_Exec->ckptMeta.fs, 1, MPI_LONG, fileSizes, 1, MPI_LONG,
      FTI_Exec->groupComm);
 
@@ -858,7 +858,7 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
         FTI_Exec->ckptMeta.pfs = fileSizes[ptnerGroupRank];
     }
 
-    int32_t mfs = 0;  // Max file size in group
+    long mfs = 0;  // Max file size in group
     int i;
     for (i = 0; i < FTI_Topo->groupSize; i++) {
         if (fileSizes[i] > mfs) {
@@ -900,11 +900,11 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
     // Every process has the same number of protected variables
 
     int* allVarIDs = NULL;
-    int32_t* allVarSizes = NULL;
-    int32_t *allVarPositions = NULL;
+    long* allVarSizes = NULL;
+    long *allVarPositions = NULL;
 
     // for posix dcp
-    uint32_t* allLayerSizes = NULL;
+    unsigned long* allLayerSizes = NULL;
     char* allLayerHashes = NULL;
     char* allCharIds = NULL;
 
@@ -913,13 +913,13 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
 
     if (FTI_Topo->groupRank == 0) {
         allVarIDs = talloc(int, FTI_Topo->groupSize * FTI_Exec->nbVar);
-        allVarSizes = talloc(int32_t, FTI_Topo->groupSize * FTI_Exec->nbVar);
-        allVarPositions = talloc(int32_t,
+        allVarSizes = talloc(long, FTI_Topo->groupSize * FTI_Exec->nbVar);
+        allVarPositions = talloc(long,
          FTI_Topo->groupSize * FTI_Exec->nbVar);
         allCharIds = (char *)malloc(sizeof(char)*FTI_BUFS*
           FTI_Exec->nbVar*FTI_Topo->groupSize);
         if (FTI_Ckpt[FTI_Exec->ckptMeta.level].isDcp) {
-            allLayerSizes = talloc(uint32_t,
+            allLayerSizes = talloc(unsigned long,
              FTI_Topo->groupSize * nbLayer);
             allLayerHashes = talloc(char,
              FTI_Topo->groupSize * nbLayer * MD5_DIGEST_STRING_LENGTH);
@@ -927,8 +927,8 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
     }
 
     int* myVarIDs = talloc(int, FTI_Exec->nbVar);
-    int32_t* myVarSizes = talloc(int32_t, FTI_Exec->nbVar);
-    int32_t* myVarPositions = talloc(int32_t, FTI_Exec->nbVar);
+    long* myVarSizes = talloc(long, FTI_Exec->nbVar);
+    long* myVarPositions = talloc(long, FTI_Exec->nbVar);
     char *ArrayOfStrings = (char *)malloc(FTI_Exec->nbVar *
      sizeof(char*) *FTI_BUFS);
 
@@ -997,3 +997,4 @@ int FTI_CreateMetadata(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
 
     return FTI_SCES;
 }
+
